pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audiohaze"
dynamic = ["version"]
description = 'This Is A Simple GUI Application To Manipulate Audio Files By Modifying The Audio Wave Form And Display It Using Python'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = ["GUI", "Tkinter", "Audio", "App", "Audio Editing", "Audio Processing"]
authors = [
    { name = "shalabycr7", email = "abdoshalaby.dev@gmail.com" },
    { name = "Abdulrahman El Sharabasy", email = "abdo.make73@gmail.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "numpy~ = 1.24.2",
    "scipy~ = 1.10.0",
    "matplotlib~ = 3.7.0",
    "pyttsx3~ = 2.90",
    "ttkbootstrap~ = 1.10.1",
    "Pillow~ = 9.5.0",
    "pydub~ = 0.25.1",
    "sounddevice~ = 0.4.6",
    "soundfile~ = 0.12.1",
]

[project.urls]
Documentation = "https://github.com/shalabycr7/audiohaze#readme"
Issues = "https://github.com/shalabycr7/audiohaze/issues"
Source = "https://github.com/shalabycr7/audiohaze"

[tool.hatch.version]
path = "src/audiohaze/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov = [
    "test-cov",
    "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.7", "3.8", "3.9", "3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
    "black>=23.1.0",
    "mypy>=1.0.0",
    "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/audiohaze tests}"
style = [
    "ruff {args:.}",
    "black --check --diff {args:.}",
]
fmt = [
    "black {args:.}",
    "ruff --fix {args:.}",
    "style",
]
all = [
    "style",
    "typing",
]

[tool.black]
target-version = ["py37"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py37"
line-length = 120
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Allow boolean positional values in function calls, like `dict.get(... True)`
    "FBT003",
    # Ignore checks for possible passwords
    "S105", "S106", "S107",
    # Ignore complexity
    "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
    # Don't touch unused imports
    "F401",
]

[tool.ruff.isort]
known-first-party = ["audiohaze"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["audiohaze", "tests"]
branch = true
parallel = true
omit = [
    "src/audiohaze/__about__.py",
]

[tool.coverage.paths]
audiohaze = ["src/audiohaze", "*/audiohaze/src/audiohaze"]
tests = ["tests", "*/audiohaze/tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
